requirement.txt
fastapi
uvicorn
motor
python-dotenv
passlib[bcrypt]
pyjwt
groq
pydantic